backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic==2.10.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
